# backend/openapi_test.json
{"openapi": "3.1.0", "info": {"title": "TeXTile API", "version": "1.0.0"}, "paths": {"/api/v1/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "operationId": "register_api_v1_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRegister"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLogin"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Get Me", "operationId": "get_me_api_v1_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/profile": {"get": {"tags": ["profile"], "summary": "Get Profile", "operationId": "get_profile_api_v1_profile_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Output"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["profile"], "summary": "Update Profile", "operationId": "update_profile_api_v1_profile_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Input"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Output"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/profile/import": {"post": {"tags": ["profile"], "summary": "Import Resume", "description": "Uploads a PDF or LaTeX file, extracts text, and uses an LLM to build a Profile object.", "operationId": "import_resume_api_v1_profile_import_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_import_resume_api_v1_profile_import_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Output"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/profile/experiences/{exp_id}/bullets/{bullet_id}/tags": {"patch": {"tags": ["profile"], "summary": "Update Bullet Tags", "operationId": "update_bullet_tags_api_v1_profile_experiences__exp_id__bullets__bullet_id__tags_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "exp_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Exp Id"}}, {"name": "bullet_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Bullet Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "array", "items": {"type": "string"}, "title": "Tags"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Output"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/profile/reorder": {"patch": {"tags": ["profile"], "summary": "Reorder Items", "operationId": "reorder_items_api_v1_profile_reorder_patch", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "section", "in": "query", "required": true, "schema": {"type": "string", "title": "Section"}}, {"name": "parent_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Parent Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "array", "items": {"type": "string"}, "title": "Ordered Ids"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Profile-Output"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/variants/{variant_id}/render": {"post": {"tags": ["variants"], "summary": "Render Variant", "operationId": "render_variant_api_v1_variants__variant_id__render_post", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "variant_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Variant Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/templates": {"get": {"tags": ["templates"], "summary": "Get All Templates", "description": "List all available LaTeX templates and their configurations.", "operationId": "get_all_templates_api_v1_templates_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/TemplateConfig"}, "type": "array", "title": "Response Get All Templates Api V1 Templates Get"}}}}}}}, "/api/v1/templates/{template_id}": {"get": {"tags": ["templates"], "summary": "Get Single Template", "description": "Get the configuration for a specific template.", "operationId": "get_single_template_api_v1_templates__template_id__get", "parameters": [{"name": "template_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Template Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TemplateConfig"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/templates/{template_id}/thumbnail": {"get": {"tags": ["templates"], "summary": "Get Template Thumbnail", "description": "Generates a PNG thumbnail for the requested template.\n\nRenders with dummy user data to show what the template looks like.", "operationId": "get_template_thumbnail_api_v1_templates__template_id__thumbnail_get", "parameters": [{"name": "template_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Template Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/health": {"get": {"summary": "Health", "operationId": "health_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_import_resume_api_v1_profile_import_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_import_resume_api_v1_profile_import_post"}, "Bullet": {"properties": {"id": {"type": "string", "title": "Id"}, "text": {"type": "string", "title": "Text"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}, "order": {"type": "integer", "title": "Order", "default": 0}}, "type": "object", "required": ["text"], "title": "Bullet"}, "Education": {"properties": {"id": {"type": "string", "title": "Id"}, "institution": {"type": "string", "title": "Institution"}, "degree": {"type": "string", "title": "Degree"}, "location": {"type": "string", "title": "Location"}, "end": {"type": "string", "title": "End"}, "gpa": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Gpa"}, "coursework": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Coursework"}, "awards": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Awards"}}, "type": "object", "required": ["institution", "degree", "location", "end"], "title": "Education"}, "Experience": {"properties": {"id": {"type": "string", "title": "Id"}, "company": {"type": "string", "title": "Company"}, "role": {"type": "string", "title": "Role"}, "location": {"type": "string", "title": "Location"}, "start": {"type": "string", "title": "Start"}, "end": {"type": "string", "title": "End"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}, "bullets": {"items": {"$ref": "#/components/schemas/Bullet"}, "type": "array", "title": "Bullets", "default": []}}, "type": "object", "required": ["company", "role", "location", "start", "end"], "title": "Experience"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Profile-Input": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "name": {"type": "string", "title": "Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "linkedin": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Linkedin"}, "github": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Github"}, "website": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Website"}, "experiences": {"items": {"$ref": "#/components/schemas/Experience"}, "type": "array", "title": "Experiences", "default": []}, "projects": {"items": {"$ref": "#/components/schemas/Project"}, "type": "array", "title": "Projects", "default": []}, "education": {"items": {"$ref": "#/components/schemas/Education"}, "type": "array", "title": "Education", "default": []}, "skill_categories": {"items": {"$ref": "#/components/schemas/SkillCategory"}, "type": "array", "title": "Skill Categories", "default": []}}, "type": "object", "required": ["user_id", "name", "email", "phone"], "title": "Profile"}, "Profile-Output": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "name": {"type": "string", "title": "Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "linkedin": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Linkedin"}, "github": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Github"}, "website": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Website"}, "experiences": {"items": {"$ref": "#/components/schemas/Experience"}, "type": "array", "title": "Experiences", "default": []}, "projects": {"items": {"$ref": "#/components/schemas/Project"}, "type": "array", "title": "Projects", "default": []}, "education": {"items": {"$ref": "#/components/schemas/Education"}, "type": "array", "title": "Education", "default": []}, "skill_categories": {"items": {"$ref": "#/components/schemas/SkillCategory"}, "type": "array", "title": "Skill Categories", "default": []}}, "type": "object", "required": ["user_id", "name", "email", "phone"], "title": "Profile"}, "Project": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "tech_stack": {"type": "string", "title": "Tech Stack"}, "date": {"type": "string", "title": "Date"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}, "bullets": {"items": {"$ref": "#/components/schemas/Bullet"}, "type": "array", "title": "Bullets", "default": []}}, "type": "object", "required": ["name", "tech_stack", "date"], "title": "Project"}, "SkillCategory": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "items": {"items": {"type": "string"}, "type": "array", "title": "Items", "default": []}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}}, "type": "object", "required": ["name"], "title": "SkillCategory"}, "TemplateConfig": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags"}, "max_experiences": {"type": "integer", "title": "Max Experiences", "default": 4}, "max_projects": {"type": "integer", "title": "Max Projects", "default": 4}, "max_skill_categories": {"type": "integer", "title": "Max Skill Categories", "default": 4}, "supports": {"$ref": "#/components/schemas/TemplateSupports", "default": {"summary": false, "photo": false, "color_accent": false, "two_column_skills": false}}, "section_order_locked": {"type": "boolean", "title": "Section Order Locked", "default": false}, "required_sections": {"items": {"type": "string"}, "type": "array", "title": "Required Sections", "default": ["Education", "Experience"]}}, "type": "object", "required": ["id", "name", "tags"], "title": "TemplateConfig"}, "TemplateSupports": {"properties": {"summary": {"type": "boolean", "title": "Summary", "default": false}, "photo": {"type": "boolean", "title": "Photo", "default": false}, "color_accent": {"type": "boolean", "title": "Color Accent", "default": false}, "two_column_skills": {"type": "boolean", "title": "Two Column Skills", "default": false}}, "type": "object", "title": "TemplateSupports"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "TokenResponse"}, "UserLogin": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "UserLogin"}, "UserRegister": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "name", "password"], "title": "UserRegister"}, "UserResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "email", "name", "created_at"], "title": "UserResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/v1/auth/login"}}}}}}